Renaming ignores to be unique to platform/browser combinations
